Fixes #1 - refactor constant classes
- Added a "constants" package and moved constants classes there
- renamed "Separators" class to "Separator" for proper naming convention
