+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.joshualeepenn</groupId>
     <artifactId>utilities</artifactId>
-    <version>0.0.1</version>
+    <version>0.0.2</version>
 
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
